+++ simulation/src/main/java/com/palantir/dialogue/core/SimulationUtils.java
@@ -16,41 +16,12 @@
 
 package com.palantir.dialogue.core;
 
-import com.google.common.collect.ListMultimap;
 import com.palantir.dialogue.Endpoint;
 import com.palantir.dialogue.HttpMethod;
-import com.palantir.dialogue.Response;
 import com.palantir.dialogue.UrlBuilder;
-import com.palantir.tritium.metrics.registry.TaggedMetricRegistry;
-import java.io.InputStream;
 import java.util.Map;
 
 final class SimulationUtils {
-
-    public static Response wrapWithCloseInstrumentation(Response delegate, TaggedMetricRegistry registry) {
-        return new Response() {
-            @Override
-            public InputStream body() {
-                return delegate.body();
-            }
-
-            @Override
-            public int code() {
-                return delegate.code();
-            }
-
-            @Override
-            public ListMultimap<String, String> headers() {
-                return delegate.headers();
-            }
-
-            @Override
-            public void close() {
-                MetricNames.responseClose(registry).inc();
-                delegate.close();
-            }
-        };
-    }
 
     static final String CHANNEL_NAME = "test-channel";
     static final String SERVICE_NAME = "svc";
